Finalize Metric naming
* Update namespace for metrics
* Move metrics to separate Maven component
* Add 5s timeout to ContainerInsights HTTP client
* Wrap log group and context discovery in exception handling
* Fix Javadoc formatting
