+++ aws-xray-recorder-metrics/src/main/java/com/amazonaws/xray/metrics/NoOpMetricEmitter.java
