+++ aws-xray-recorder-metrics/src/main/java/com/amazonaws/xray/metrics/StdoutMetricEmitter.java
