+++ aws-xray-recorder-metrics/src/test/java/com/amazonaws/xray/config/MetricsDaemonConfigurationTest.java
