+++ pom.xml
@@ -42,6 +42,7 @@
     <module>aws-xray-recorder-sdk-aws-sdk-v2-instrumentor</module>
     <module>aws-xray-recorder-sdk-slf4j</module>
     <module>aws-xray-recorder-sdk-log4j</module>
+    <module>aws-xray-recorder-metrics</module>
   </modules>
   <distributionManagement>
     <repository>
